Follow the pattern from others
# Conflicts:
#	src/Controls/tests/DeviceTests/Elements/HybridWebView/HybridWebViewTests.cs
