+++ src/Core/src/Handlers/HybridWebView/HybridWebViewHandler.Android.cs
@@ -15,8 +15,6 @@ namespace Microsoft.Maui.Handlers
 
 		protected override AWebView CreatePlatformView()
 		{
-			var customSettings = VirtualView.InitializingWebView() as WebViewInitializingEventArgs;
-
 			var platformView = new MauiHybridWebView(this, Context!)
 			{
 				LayoutParameters = new LayoutParams(LayoutParams.MatchParent, LayoutParams.MatchParent)
@@ -34,13 +32,16 @@ namespace Microsoft.Maui.Handlers
 
 			platformView.Settings.JavaScriptEnabled = true;
 
-			// Allow additional custom settings to be applied through the WebViewInitializingEventArgs Class
-			customSettings?.AdditionalSettings?.Invoke(platformView.Settings);
-
 			_javaScriptInterface = new HybridWebViewJavaScriptInterface(this);
 			platformView.AddJavascriptInterface(_javaScriptInterface, HybridWebViewHostJsName);
 
-			VirtualView.InitializedWebView(platformView);
+			// Invoke the WebViewInitializing event to allow custom configuration of the web view
+			var initializingArgs = new WebViewInitializationStartedEventArgs(platformView, platformView.Settings);
+			VirtualView.WebViewInitializationStarted(initializingArgs);
+
+			// Invoke the WebViewInitialized event to signal that the web view has been initialized
+			var initializedArgs = new WebViewInitializationCompletedEventArgs(platformView);
+			VirtualView?.WebViewInitializationCompleted(initializedArgs);
 
 			return platformView;
 		}
